+++ lib/net_buf/buf.c
@@ -79,7 +79,7 @@ static inline struct net_buf *pool_get_uninit(struct net_buf_pool *pool,
 
 	buf = (struct net_buf *)(((uint8_t *)pool->__bufs) + byte_offset);
 
-	buf->pool_id = pool_id(pool);
+	buf->pool_id = (uint8_t)pool_id(pool);
 	buf->user_data_size = pool->user_data_size;
 
 	return buf;
@@ -331,7 +331,7 @@ success:
 	buf->ref   = 1U;
 	buf->flags = 0U;
 	buf->frags = NULL;
-	buf->size  = size;
+	buf->size  = (uint16_t)size;
 	memset(buf->user_data, 0, buf->user_data_size);
 	net_buf_reset(buf);
 
@@ -508,7 +508,7 @@ struct net_buf *net_buf_clone(struct net_buf *buf, k_timeout_t timeout)
 			return NULL;
 		}
 
-		clone->size = size;
+		clone->size = (uint16_t)size;
 		clone->data = clone->__buf + net_buf_headroom(buf);
 		net_buf_add_mem(clone, buf->data, buf->len);
 	}
@@ -656,7 +656,7 @@ size_t net_buf_append_bytes(struct net_buf *buf, size_t len,
 	size_t max_size;
 
 	do {
-		uint16_t count = MIN(len, net_buf_tailroom(frag));
+		size_t count = MIN(len, net_buf_tailroom(frag));
 
 		net_buf_add_mem(frag, value8, count);
 		len -= count;
